+++ include/beast/http.hpp
@@ -11,8 +11,9 @@
 #include <beast/config.hpp>
 
 #include <beast/http/basic_parser.hpp>
-#include <beast/http/chunk_encode.hpp>
+#include <beast/http/buffer_body.hpp>
 #include <beast/http/dynamic_body.hpp>
+#include <beast/http/empty_body.hpp>
 #include <beast/http/error.hpp>
 #include <beast/http/fields.hpp>
 #include <beast/http/header_parser.hpp>