+++ test/http/error.cpp
@@ -38,6 +38,7 @@ public:
     {
         check("http", error::end_of_stream);
         check("http", error::partial_message);
+        check("http", error::need_more);
         check("http", error::buffer_overflow);
         check("http", error::bad_line_ending);
         check("http", error::bad_method);